Move generic push functions to Push.hs
